extracted ZaberDriver
